+++ activities/Abacus.activity/js/onecolumnabacus.js
@@ -328,6 +328,11 @@ function OneColumnAbacus(stage,rods,number,base,colours,startvalue,schety,fracti
 			} else {
 				this.answertext.text = "";
 			}
+			if((this.answertext.text).length >= 40) {
+				this.answertext.font =  parseInt(((this.blockWidth / (this.answertext.text).length)*40)) + "px Arial";
+			} else {
+				this.answertext.font =  (this.blockWidth).toString()+"px Arial";
+			}
 		}
 	}
 
@@ -367,7 +372,13 @@ function OneColumnAbacus(stage,rods,number,base,colours,startvalue,schety,fracti
 			} else {
 				this.answertext.text = "";
 			}
+			if((this.answertext.text).length >= 40) {
+				this.answertext.font =  parseInt(((this.blockWidth / (this.answertext.text).length)*40)) + "px Arial";
+			} else {
+				this.answertext.font =  (this.blockWidth).toString()+"px Arial";
+			}
 		}
+		
 	}
 
 	this.init = function(){