+++ CHANGELOG.md
@@ -9,10 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Desktop version for GNU Linux, MacOS and Windows using electron
 - Ebook Reader activity
 - Exerciser activity
+- Sprint Math activity
 - Add a sort palette in Journal
 - Allow action on multiple items in Journal
 - Copy from Journal to device now available on all platforms
 - Copy from device to Journal feature
+- Help tutorial on initial screen
+- Help on activities in list view
 
 ### Changed
 - Portuguese localization
@@ -24,6 +27,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update Lunar ephemerides phase for 2018 to 2024 in Moon activity #219
 - DOM-based XSS vulnerability in the MarkDown Editor Activity #158
 - DOM-based XSS vulnerability in the Get Things Done Activity #160
+- Save Stopwatch activity context in Journal #201
+- Flip activity don't detect end of game #215
+- Align buttons better in tutorial #135
+- ChatPrototype activity does not sanitize input #225
+- Fototoon close icon missing in small screens #213
+- Save Clock activity settings in Journal #204
+- Add number of play in Flip activity #234
+- Save Blockrain activity context in Journal #203
+- Activity Idea: Math Quiz #187
+- Abacus does not display solution with big numbers #140
+- Change background color for Speak activity #245
+- Add the sleeping face into the Speak activity #246 
+
+### Removed
+- Remove Cordova activity
+- Remove Welcome Web activity
 
 
 ## [1.0.1] - 2018-07-07